+++ .gitignore
@@ -1,3 +1,6 @@
+# local testing outputs
+tests/wavanim*
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
